+++ docs/CHANGELOG.md
@@ -23,6 +23,10 @@ Entries inside each section should be ordered by type:
 * [Removed] Drop support for Python 3.9 (end-of-life); `quilt3` now requires Python >= 3.10 ([#4941](https://github.com/quiltdata/quilt/pull/4941))
 * [Fixed] `quilt3.admin.buckets.list` no longer raises `TypeError` when its type hints are introspected on Python 3.14 ([#4940](https://github.com/quiltdata/quilt/pull/4940))
 
+### CLI
+
+* [Changed] `quilt3 login` no longer echoes the pasted code to the terminal (it is a long-lived credential) ([#XXXX](https://github.com/quiltdata/quilt/pull/XXXX))
+
 ## 7.3.0 - 2026-04-07
 
 ### Python API